BackgroundStyle now also has support for preview images.
